// sourcecode/BTrak.Packages/snova-app-builder-creation-components/app-builder-module-components/app-builder-components/src/lib/builder/components/genericform/models/createGenericForm.ts
export class CreateGenericForm {
    Id: string;
    FormTypeId: string;
    FormType: string;
    formTypeName: string;
    formName: string;
    workflowTrigger: string;
    formJson: string;
    createdDateTime: Date;
    CreatedByUserId: string;
    UpdatedDateTime: Date;
    UpdatedByUserId: string;
    IsArchived: boolean;
    ArchivedDateTime: Date;
    ArchivedByUserId: string;
    IsDeleted: boolean;
    TimeStamp: any;
    roleIds : string;
    formKeys: string;
    profileImage: string;
    fullName: string;
    statusReportingConfigurationId: string;
    isSubmitted: boolean;
    statusReportingConfigurationOptionId: string;
    isAbleToLogin: boolean;
    customApplicationNames: string[];
    fields: any;
    dataSourceId: any;
    isInludeTemplateForms: boolean;
    companiesList: any;
    viewFormRoleIds: any;
    editFormRoleIds: any;
}